refactor(useCartItems): 테스트 케이스 명세 변경 및 파일 위치 수정
