+++ src/hooks/useCartItems.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { QUERY_KEYS } from '../../constants/queryKeys';
-import { fetchCartItems } from '../../api/cartItems';
+import { QUERY_KEYS } from '../constants/queryKeys';
+import { fetchCartItems } from '../api/cartItems';
 
 const useCartItems = () => {
   return useQuery({
